test: move unit tests for Ethereum client block retrieval and retry logic to client module
